refac: rename and move some files for better understanding
