+++ src/main/scala/api/bucket/models/Get.scala
@@ -1,22 +1,26 @@
 package pedro.goncalves
-package api.bucket
+package api.bucket.models
+
+import api.bucket.{Bucket, Buckets}
 
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
 import spray.json.{DefaultJsonProtocol, RootJsonFormat}
 
 
+
 final case class Bucket(
-                  name: String
-                       )
+                  id:String,
+                  name:String
+                          )
 
 final case class Buckets(
                   buckets:Seq[Bucket]
                         )
 
 
-trait BucketJsonSupport extends SprayJsonSupport with DefaultJsonProtocol {
+trait GetJsonSupport extends SprayJsonSupport with DefaultJsonProtocol {
 
-  implicit val bucketFormat: RootJsonFormat[Bucket] = jsonFormat1(Bucket.apply)
+  implicit val getFormat: RootJsonFormat[Bucket] = jsonFormat2(Bucket.apply)
 
   implicit val bucketsFormat: RootJsonFormat[Buckets] = jsonFormat1(Buckets.apply)
